Revert "Синхронизация мини плеера и обычного"
This reverts commit 222c7290409cb7002f121bc2f2591c5cb73d2342.
