+++ MyMusicApp/Modules/Song/Controllers/SongPageViewController.swift
@@ -7,27 +7,13 @@
 
 import UIKit
 
-struct SongInfo {
-    var model: Entry
-    var image: UIImage
-}
-
 final class SongPageViewController: UIPageViewController {
-    
-    var songInfo: SongInfo? {
-        didSet {
-            if let songInfo = songInfo {
-                songVC.configurateView(model: songInfo.model, image: songInfo.image)
-                albumVC.configurateView(model: songInfo.model, image: songInfo.image)
-            }
-        }
-    }
-    
-    let songVC = SongViewController()
-    let albumVC = AlbumViewController()
     
     lazy var controllers: [UIViewController] = {
         var controllers = [UIViewController]()
+        
+        let songVC = SongViewControllers()
+        let albumVC = AlbumViewController()
         
         controllers.append(songVC)
         controllers.append(albumVC)
@@ -39,7 +25,6 @@ final class SongPageViewController: UIPageViewController {
         super.init(transitionStyle: .scroll, navigationOrientation: navigationOrientation)
         setViewControllers([controllers[0]], direction: .forward, animated: true)
         dataSource = self
-        delegate = self
     }
     
     required init?(coder: NSCoder) {
@@ -67,11 +52,5 @@ extension SongPageViewController: UIPageViewControllerDataSource {
         }
         
         return nil
-    }
-}
-
-extension SongPageViewController: UIPageViewControllerDelegate {
-    func pageViewController(_ pageViewController: UIPageViewController, didFinishAnimating finished: Bool, previousViewControllers: [UIViewController], transitionCompleted completed: Bool) {
-        
     }
 }
