--- MyMusicApp/Modules/Song/Protocols/SongViewControllerProtocol.swift
@@ -1,17 +1,0 @@
-//
-//  SongViewControllerProtocol.swift
-//  MyMusicApp
-//
-//  Created by Евгений on 21.06.2023.
-//
-
-import UIKit
-
-protocol SongViewControllerProtocol: AnyObject {
-    func updateButtonImage(isPlay: Bool)
-    func updateCurrentTimeLabel(duration: Int)
-    func updateTotalDuration(duration: Float)
-    func updateSlider(value: Float)
-    func configurateView(model: Entry, image: UIImage)
-    func setDurationTime()
-}
